feat: updates solution to 4 series
