+++ .github/workflows/EssentialsPlugins-builds-caller.yml
@@ -1,0 +1,30 @@
+name: Build Essentials Plugin
+
+on:
+  push:
+    branches-ignore:
+      - 'robot-docs'
+jobs:
+  checkCommitMessage:
+    uses: PepperDash/workflow-templates/.github/workflows/essentialsplugins-checkCommitMessage.yml@development
+    secrets: inherit
+  call-update-readme:
+    uses: PepperDash/readme-automation/.github/workflows/update-readme.yml@main
+    secrets: inherit
+    needs: checkCommitMessage
+    with:
+      target-branch: ${{ github.ref_name }}
+  getVersion:
+    uses: PepperDash/workflow-templates/.github/workflows/essentialsplugins-getversion.yml@development
+    secrets: inherit
+    needs: checkCommitMessage
+  build-4Series:
+    uses: PepperDash/workflow-templates/.github/workflows/essentialsplugins-4Series-builds.yml@development
+    secrets: inherit
+    needs: getVersion
+    if: needs.getVersion.outputs.newVersion == 'true'
+    with:
+      newVersion: ${{ needs.getVersion.outputs.newVersion }}
+      version: ${{ needs.getVersion.outputs.version }}
+      tag: ${{ needs.getVersion.outputs.tag }}
+      channel: ${{ needs.getVersion.outputs.channel }}
