--- DynFusionEPI/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("EssentialsPluginTemplateEpi")]
-[assembly: AssemblyCompany("PepperDash")]
-[assembly: AssemblyProduct("EssentialsPluginTemplateEpi")]
-[assembly: AssemblyCopyright("Copyright ©  2023")]
-[assembly: AssemblyVersion("1.0.0.*")]
-
