--- DynFusionEPI/SchedulingConfig.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Crestron.SimplSharp;
-using PepperDash.Essentials.Core;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace DynFusion
-{
-	public class SchedulingConfig
-	{
-		public string DynFusionKey {get; set;}
-	}
-
-
-}
