+++ epi-crestron-fusion.4series.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "epi-crestron-fusion.4series", "src\epi-crestron-fusion.4series.csproj", "{B28AA204-985A-4730-974C-FEBA5E7894C6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B28AA204-985A-4730-974C-FEBA5E7894C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B28AA204-985A-4730-974C-FEBA5E7894C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B28AA204-985A-4730-974C-FEBA5E7894C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B28AA204-985A-4730-974C-FEBA5E7894C6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C239DD33-FA40-42F8-8B96-9689D52B2780}
+	EndGlobalSection
+EndGlobal
