+++ src/Config/AssetsClass.cs
