+++ src/Config/CustomProperties.cs
