+++ src/Config/DynFusionConfigObjectTemplate.cs
