+++ src/Config/FusionCustomProperty.cs
