+++ src/Config/FusionOccupancyAsset.cs
