+++ src/Config/FusionStaticAssetConfig.cs
