+++ src/DynFusionStaticAsset.cs
