+++ src/DynFusionStaticAssetJoinMap.cs
