+++ src/epi-crestron-fusion.4series.csproj
@@ -1,0 +1,36 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<ProjectType>ProgramLibrary</ProjectType>
+	</PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net472</TargetFramework>
+		<RootNamespace>Pepperdash.Essentials.Plugins.Crestron.Fusion</RootNamespace>
+		<Deterministic>false</Deterministic>
+		<AssemblyTitle>EPI.Crestron.Fusion</AssemblyTitle>
+		<Company>PepperDash Technologies</Company>
+		<Description>This software is a plugin designed to work as a part of PepperDash Essentials for Crestron control processors. This plugin allows for integration of Crestron Fusion.</Description>
+		<Copyright>Copyright 2024</Copyright>
+		<Version>1.0.0-local</Version>
+		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
+		<InformationalVersion>$(Version)</InformationalVersion>
+		<OutputPath>4Series\bin\$(Configuration)\</OutputPath>
+		<Authors>PepperDash Technologies</Authors>
+		<PackageId>Pepperdash.Essentials.Plugins.Crestron.Fusion</PackageId>
+		<PackageProjectUrl>https://github.com/PepperDash/epi-crestron-fusion</PackageProjectUrl>
+		<PackageTags>crestron 4series fusion</PackageTags>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+		<DefineConstants>$(DefineConstants);SERIES4</DefineConstants>
+	</PropertyGroup>
+	<ItemGroup>
+		<Compile Remove="Properties\**" />
+		<EmbeddedResource Remove="Properties\**" />
+		<None Remove="Properties\**" />
+	</ItemGroup>
+	<ItemGroup>
+		<None Remove="PDTDynFusionEPI.projectinfo" />
+	</ItemGroup>
+	<ItemGroup>
+	  <PackageReference Include="PepperDashEssentials" Version="2.0.0-alpha-2600" />
+	</ItemGroup>
+</Project>
